fix: update link and extra config files
