+++ LICENSE
@@ -1,6 +1,6 @@
 The MIT License
 
-Copyright (c) 2015 - present NHN Entertainment Corp.
+Copyright (c) 2017 - present NHN Entertainment Corp.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
