--- examples/example00.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="utf-8" />
-<title>Toast UI Colorpicker</title>
-<link rel="stylesheet" type="text/css" href="../dist/tui-color-picker.min.css" />
-<style>
-    #colorpicker {
-        width: 166px;
-        border: 1px solid #ccc;
-        border-radius: 3px;
-        padding: 7px;
-    }
-</style>
-</head>
-<body>
-<h1>Toast UI Colorpicker</h1>
-<div class="wrap">
-    <div id="colorpicker"></div>
-</div>
-
-<script src="https://cdn.rawgit.com/nhnent/tui.code-snippet/v1.2.5/dist/tui-code-snippet.js"></script>
-<script src="../dist/tui-color-picker.js"></script>
-
-<script>
-(function() {
-    var colorpicker = tui.colorPicker.create({
-        container: document.getElementById('colorpicker')
-    });
-
-    colorpicker.on('selectColor', function(obj) {
-        console.log(obj);
-    });
-})();
-</script>
-</body>
-</html>
