+++ examples/example01-basic.html
@@ -1,0 +1,36 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8" />
+        <title>1. Basic</title>
+        <link href="../dist/tui-color-picker.css" rel="stylesheet" />
+        <style type="text/css">
+            #color-picker {
+                width: 166px;
+                border: 1px solid #ccc;
+                border-radius: 3px;
+                padding: 7px;
+            }
+        </style>
+    </head>
+
+    <body>
+        <div class="code-html">
+            <div id="color-picker"></div>
+        </div>
+
+        <script src="https://cdn.rawgit.com/nhnent/tui.code-snippet/v1.2.5/dist/tui-code-snippet.js"></script>
+        <script src="../dist/tui-color-picker.js"></script>
+        <script class="code-js">
+
+        var colorpicker = tui.colorPicker.create({
+            container: document.getElementById('color-picker')
+        });
+
+        colorpicker.on('selectColor', function(obj) {
+            console.log(obj);
+        });
+
+        </script>
+    </body>
+</html>
