+++ examples/examples.json
@@ -1,5 +1,5 @@
 {
-  "sample1": {
-    "title": "ColorPicker - Basic"
-  }
+    "example01-basic": {
+        "title": "1. Basic"
+    }
 }
